+++ platform/ui/src/index.js
@@ -1,7 +1,7 @@
 import {
+  Checkbox,
   CineDialog,
   DownloadDialog,
-  ExampleDropTarget,
   LayoutButton,
   LayoutChooser,
   MeasurementTable,
@@ -37,42 +37,43 @@ import {
 } from './elements';
 
 // Alias this for now as not all dependents are using strict versioning
+import { DropdownMenu as Dropdown, Range, Select } from './elements/form';
 import ExpandableToolMenu from './viewer/ExpandableToolMenu.js';
-import LayoutManager from './LayoutChooser/LayoutManager.js';
-import LayoutPanelDropTarget from './LayoutChooser/LayoutPanelDropTarget.js';
 import PlayClipButton from './viewer/PlayClipButton.js';
 import { ScrollableArea } from './ScrollableArea/ScrollableArea.js';
 import Toolbar from './viewer/Toolbar.js';
 import ToolbarButton from './viewer/ToolbarButton.js';
 import ViewerbaseDragDropContext from './utils/viewerbaseDragDropContext.js';
+import SnackbarProvider, {
+  useSnackbarContext,
+  withSnackbar,
+} from './utils/SnackbarProvider';
 
 export {
   // Elements
   ICONS,
-  Icon,
+  //
+  Checkbox,
   Dropdown,
-  Select,
   Label,
-  Range,
-  TextInput,
   TextArea,
-  // Components
+  TextInput,
   CineDialog,
   DownloadDialog,
   ExpandableToolMenu,
-  ExampleDropTarget,
+  Icon,
   LayoutButton,
   LayoutChooser,
-  LayoutManager,
-  LayoutPanelDropTarget,
   MeasurementTable,
   MeasurementTableItem,
   Overlay,
   OverlayTrigger,
   PlayClipButton,
   QuickSwitch,
+  Range,
   RoundedButtonGroup,
   ScrollableArea,
+  Select,
   SelectTree,
   SimpleDialog,
   StudyBrowser,
@@ -88,4 +89,7 @@ export {
   UserPreferences,
   UserPreferencesModal,
   ViewerbaseDragDropContext,
+  SnackbarProvider,
+  useSnackbarContext,
+  withSnackbar,
 };