+++ platform/viewer/.webpack/webpack.pwa.js
@@ -32,26 +32,10 @@ module.exports = (env, argv) => {
   const hasProxy = PROXY_TARGET && PROXY_DOMAIN;
 
   const mergedConfig = merge(baseConfig, {
-    devtool: isProdBuild ? 'source-map' : 'cheap-module-eval-source-map',
     output: {
       path: DIST_DIR,
       filename: isProdBuild ? '[name].bundle.[chunkhash].js' : '[name].js',
       publicPath: PUBLIC_URL, // Used by HtmlWebPackPlugin for asset prefix
-    },
-    stats: {
-      colors: true,
-      hash: true,
-      timings: true,
-      assets: true,
-      chunks: false,
-      chunkModules: false,
-      modules: false,
-      children: false,
-      warnings: true,
-    },
-    optimization: {
-      minimize: isProdBuild,
-      sideEffects: true,
     },
     module: {
       rules: [...extractStyleChunksRule(isProdBuild)],
@@ -71,6 +55,12 @@ module.exports = (env, argv) => {
           // Ignore our configuration files
           ignore: ['config/*', 'html-templates/*', '.DS_Store'],
         },
+        // Short term solution to make sure GCloud config is available in output
+        // for our docker implementation
+        {
+          from: `${PUBLIC_DIR}/config/google.js`,
+          to: `${DIST_DIR}/google.js`,
+        },
         // Copy over and rename our target app config file
         {
           from: `${PUBLIC_DIR}/${APP_CONFIG}`,
@@ -79,8 +69,8 @@ module.exports = (env, argv) => {
       ]),
       // https://github.com/faceyspacey/extract-css-chunks-webpack-plugin#webpack-4-standalone-installation
       new ExtractCssChunksPlugin({
-        filename: '[name].css',
-        chunkFilename: '[id].css',
+        filename: isProdBuild ? '[name].[hash].css' : '[name].css',
+        chunkFilename: isProdBuild ? '[id].[hash].css' : '[id].css',
         ignoreOrder: false, // Enable to remove warnings about conflicting order
       }),
       // Generate "index.html" w/ correct includes/imports
@@ -108,6 +98,8 @@ module.exports = (env, argv) => {
       hot: true,
       open: true,
       port: 3000,
+      host: '0.0.0.0',
+      public: 'http://localhost:' + 3000,
       historyApiFallback: {
         disableDotRule: true,
       },
@@ -121,15 +113,6 @@ module.exports = (env, argv) => {
 
   if (!isProdBuild) {
     mergedConfig.plugins.push(new webpack.HotModuleReplacementPlugin());
-
-    //
-    mergedConfig.optimization.minimizer = [
-      new TerserJSPlugin({
-        sourceMap: true,
-        parallel: true,
-      }),
-      new OptimizeCSSAssetsPlugin({}),
-    ];
   }
 
   return mergedConfig;