+++ platform/viewer/src/components/Header/Header.js
@@ -101,6 +101,14 @@ class Header extends Component {
             </Link>
           )}
 
+          <span
+            className={`${
+              this.props.home ? 'header-versionInfoHome' : 'header-versionInfo'
+            }`}
+          >
+            v{process.env.VERSION_NUMBER}
+          </span>
+
           {this.props.children}
 
           {showStudyList && !this.props.home && (